src/requirements.txt
fmeval
boto3
langchain
ollama
websockets
python-dotenv
langchain-aws